Split unrollers into separate files
